payment email: correctly render templates, clean up code
