--- app/views/mailings/payment/requestmail.html.haml
@@ -1,16 +1,0 @@
--# %p
--#   = I18n.t('mailings.greeting')
--#   -# = name
-
--# %p
--#   = I18n.t('mailings.payment.open_activities')
--#   -# = activities
-
--# -# %p
--# -#   = I18n.t('mailings.payment.total_of', amount: total)
-
--# %p
--#   = I18n.t('mailings.best_regards')
-
--# %p
--#   = I18n.t('mailings.signature')
