+++ config/locales/mailings.nl.yml
@@ -67,7 +67,7 @@ nl:
       open_activities: 'Je hebt de volgende activiteiten nog niet betaald:'
       payment_instruction: Zou je dit bedrag zo snel mogelijk via het <a href="https://koala.svsticky.nl/member/payments">betalingsportaal op Koala</a> willen betalen?
       subject: Openstaande betalingen
-      total_of: Dit geeft een totaal van %{amount}
+      total_of: Dit geeft een totaal van €%{amount}
     removed: is direct verwijderd
     sent:
       'false': Mail is niet verstuurd!
